requirements.txt
certifi==2020.11.8
chardet==3.0.4
deprecation==2.1.0
eyeD3==0.9.5
filetype==1.0.7
idna==2.10
packaging==20.4
Pillow==8.0.1
pyparsing==2.4.7
requests==2.25.0
six==1.15.0
urllib3==1.26.2
youtube-dl==2020.11.24
